pyproject.toml
[tool.poetry]
name = "gptcli"
version = "0.1.0"
description = ""
authors = ["Ben Van Sleen <benvansleen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.27.2"
rich = "^13.3.2"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gptcli = "gptcli.__main__:main"
